move transformer to its own class
