+++ modules/lib/core/src/main/scala/org/corespring/platform/core/services/BaseContentService.scala
@@ -8,7 +8,6 @@ import org.corespring.platform.core.models.error.CorespringInternalError
 import com.mongodb.casbah.commons.MongoDBObject
 import org.corespring.platform.core.models.item.Item.Keys._
 import com.novus.salat.dao.SalatMongoCursor
-import scala.Some
 
 trait BaseFindAndSaveService[ContentType <: Content[ID], ID] {
 
@@ -27,11 +26,7 @@ trait BaseContentService[ContentType <: Content[ID], ID] extends BaseFindAndSave
 
   def find(query: DBObject, fields: DBObject = new BasicDBObject()): SalatMongoCursor[ContentType]
 
-  def findOneById(id: ID): Option[ContentType]
-
   def findOne(query: DBObject): Option[ContentType]
-
-  def save(i: ContentType, createNewVersion: Boolean = false)
 
   def insert(i: ContentType): Option[ID]
